+++ lithops/standalone/backends/aws_ec2/aws_ec2.py
@@ -65,7 +65,7 @@ class AWSEC2Backend:
         self.ec2_client = boto3.client(
             'ec2', aws_access_key_id=ec2_config['access_key_id'],
             aws_secret_access_key=ec2_config['secret_access_key'],
-            aws_session_token=ec2_config['session_token'],
+            aws_session_token=ec2_config.get('session_token'),
             config=client_config,
             region_name=self.region
         )
@@ -303,7 +303,7 @@ class EC2Instance:
         ec2_client = boto3.client(
             'ec2', aws_access_key_id=self.config['access_key_id'],
             aws_secret_access_key=self.config['secret_access_key'],
-            aws_session_token=self.config['session_token'],
+            aws_session_token=self.config.get('session_token', None),
             config=client_config,
             region_name=self.region
         )