+++ lithops/storage/backends/aws_s3/aws_s3.py
@@ -51,7 +51,7 @@ class S3Backend:
         self.s3_client = boto3.client(
             's3', aws_access_key_id=s3_config['access_key_id'],
             aws_secret_access_key=s3_config['secret_access_key'],
-            aws_session_token=s3_config['session_token'],
+            aws_session_token=s3_config.get('session_token'),
             config=client_config,
             endpoint_url=self.service_endpoint
         )
@@ -181,7 +181,7 @@ class S3Backend:
         max_keys_num = 1000
         for i in range(0, len(key_list), max_keys_num):
             delete_keys = {'Objects': []}
-            delete_keys['Objects'] = [{'Key': k} for k in key_list[i:i+max_keys_num]]
+            delete_keys['Objects'] = [{'Key': k} for k in key_list[i:i + max_keys_num]]
             result.append(self.s3_client.delete_objects(Bucket=bucket_name, Delete=delete_keys))
         return result